fix logout
