+++ routes/web.php
@@ -29,8 +29,8 @@ Route::group(['prefix' => 'gio-hang'], function () {
 // Customer
 Route::group(['prefix' => 'nguoi-dung'], function () {
     Route::get('dang-nhap.html', 'LoginController@index')->name('guest.user.login');
-    Route::get('dang-xuat.html', 'LoginController@logout')->name('guest.user.logout');
     Route::post('dang-nhap.html', 'LoginController@login')->name('guest.user.login');
+    Route::get('dang-xuat.html', 'LogoutController@logout')->name('guest.user.logout');
     Route::get('dang-ky.html', 'RegisterController@index')->name('guest.user.index');
 });
